+++ aks-node-controller/helpers/const.go
