+++ aks-node-controller/helpers/helper.go
