+++ aks-node-controller/helpers/helper_test.go
