+++ aks-node-controller/helpers/utils.go
