+++ aks-node-controller/helpers/utils_test.go
