+++ aks-node-controller/parser/helper.go
@@ -34,9 +34,9 @@ import (
 	"strings"
 	"text/template"
 
+	"github.com/Azure/agentbaker/aks-node-controller/helpers"
 	"github.com/Azure/agentbaker/pkg/agent"
 	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
-	"github.com/Azure/agentbaker/pkg/helpers"
 )
 
 var (
