+++ aks-node-controller/parser/helper_test.go
@@ -21,8 +21,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/Azure/agentbaker/aks-node-controller/helpers"
 	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
-	"github.com/Azure/agentbaker/pkg/helpers"
 	"github.com/Azure/go-autorest/autorest/to"
 )
