+++ aks-node-controller/parser/parser_test.go
@@ -11,10 +11,10 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/Azure/agentbaker/aks-node-controller/helpers"
 	"github.com/Azure/agentbaker/aks-node-controller/parser"
 	"github.com/Azure/agentbaker/pkg/agent/datamodel"
 	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
-	"github.com/Azure/agentbaker/pkg/helpers"
 	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
